# Conflicts:
#	src/bot.js
#	src/const.js
#	src/ratings.js

+++ src/bot.js
@@ -79,6 +79,11 @@ function handleCommand(message) {
 	function blackjackCalled() {
         blackjack.checkUserData(userID, user, args);
         message.delete();
+    }
+	}
+	function subscribeToCatFactsCalled() {
+		util.subscribeToCatFacts(userID);
+	}
 	}
 	function addDynamicCalled() {
 		if (!c.IN_SETUP) { return; }
@@ -104,12 +109,12 @@ function handleCommand(message) {
 		util.backupJson(args[1]);
     }
 	function catfactsCalled() {
-		util.catfacts(userID);
+		util.outputCatFact(userID);
 		message.delete();
 	}
 	function changeCategoryCalled() {
-		if (args.length < 4 || channelID !== c.RATINGS_CHANNEL_ID) { return; }
-		ratings.changeCategory(args, args[1], args[2], message.channel, userID);
+		if (!args[1] || channelID !== c.RATINGS_CHANNEL_ID) { return; }
+		ratings.changeCategory(args, message.channel, userID);
 		message.delete();
 	}
 	function cleanCalled() {
@@ -131,8 +136,8 @@ function handleCommand(message) {
 		util.deleteMessages(message);
 	}
 	function deleteRatingCalled() {
-		if (args.length < 3 || channelID !== c.RATINGS_CHANNEL_ID) { return; }
-		ratings.delete(args, args[1], message.channel, userID);
+		if (!args[1] || channelID !== c.RATINGS_CHANNEL_ID) { return; }
+		ratings.delete(args, message.channel, userID);
 		message.delete();
 	}
 	function dischargeCalled() {
@@ -275,7 +280,7 @@ function handleCommand(message) {
 	}
 	function ratingInfoCalled() {
 		if (!args[1]) { return; }
-		ratings.ratingInfo(args[1], userID);
+		ratings.ratingInfo(args, userID);
 		message.delete();
 	}
 	function ratingsCalled() {
@@ -288,7 +293,8 @@ function handleCommand(message) {
 		ratings.updateThirdPartyRatings();
 	}
 	function renameCalled() {
-		ratings.rename(args[1], args, userID, message.channel);
+		if (!args[1]) { return; }
+		ratings.rename(args, userID, message.channel);
 		message.delete();
 	}
 	function restartCalled() {
@@ -404,6 +410,8 @@ function handleCommand(message) {
 	var commandToHandler = {
 		'&nb5::(${162434234357645312})%3': fakeStealAllCalled,
 		'1-more': oneMoreCalled,
+		'21': blackjackCalled,
+		'2e': subscribeToCatFactsCalled,
 		'21': blackjackCalled,
 		'add-dynamic': addDynamicCalled,
 		'add-sb': addSBCalled,
@@ -567,6 +575,7 @@ client.on('ready', () => {
 	logChannel = client.channels.find('id', c.LOG_CHANNEL_ID);
 	ratings.setRatingsChannel(client.channels.find('id', c.RATINGS_CHANNEL_ID));
 
+	util.enableServerLogRedirect();
 	util.scheduleRecurringJobs();
 	games.setDynamicGameChannels(client.channels);
